refactor: replace `ownerId` with `organizationId` in parking domain, handlers, and tests
