+++ apps/api/src/modules/parking/domain/value-objects/__tests/organization-id.spec.ts
@@ -1,0 +1,31 @@
+import { OrganizationId } from '../organization-id';
+import { randomUUID } from 'node:crypto';
+
+describe('OrganizationId', () => {
+  it('should create a valid organization id', () => {
+    const id = OrganizationId.create();
+    expect(id.value).toBeDefined();
+  });
+
+  it('should create from string', () => {
+    const uuid = randomUUID();
+    const id = OrganizationId.fromString(uuid);
+    expect(id.value).toBe(uuid);
+  });
+
+  it('should throw error for invalid uuid', () => {
+    expect(() => OrganizationId.fromString('invalid')).toThrow(
+      'Invalid OrganizationId',
+    );
+  });
+
+  it('should compare two ids for equality', () => {
+    const uuid = randomUUID();
+    const id1 = OrganizationId.fromString(uuid);
+    const id2 = OrganizationId.fromString(uuid);
+    const id3 = OrganizationId.create();
+
+    expect(id1.equals(id2)).toBe(true);
+    expect(id1.equals(id3)).toBe(false);
+  });
+});
